// shared/convolo-icallback-js/types/callback_widget_schemas/Versions.ts
export enum Version {
    V30 = '20191003_30_wrong_country',
    V31 = '20191003_31_spa_enabled_flag',
    V32 = '20191112_32_open_time_selector_default',
    V33 = '20191116_33_header_height',
    V34 = '20191116_34_logoBottomGap_onDemand',
    V35 = '20191127_35_titleFontWeight',
    V36 = '20191127_36_mobileButtonPosition',
    V37 = '20200107_37_mobile_and_desktop_button_position',
    V38 = '20200121_38_show_phone_field_at_tooltip',
    V39 = '20200728_39_leadconnect_cc_and_ae_to_convolo_ai',
    V40 = '20200916_40_departments_order',
    V41 = '20201028_41_first_second_line_for_departments_and_time',
    V42 = '20220111_42_rating',
    V43 = '20220225_43_disable_external_css',
    V44 = '20220527_44_widget_form',
    V45 = '20220606_45_rating_settings',
    V46 = '20221406_46_rating_styles',
    V47 = '20222006_47_rating_const_and_lead_form_more_styles',
    V48 = '20222706_48_rating_and_form_more_styles',
    V49 = '20222806_49_mobile_templates',
    V50 = '20221208_50_show_unreacheble_rating',
    V51 = '20230211_51_set_powered_by_settings_design',
    V52 = '20230405_52_set_phone_without_dial_code',
    V53 = '20230411_52_set_rating_animation',
}

export const currentVersion = Version.V51;